requirements.txt
colorama==0.4.6
hikari==2.0.0.dev124
hikari-lightbulb==2.3.5
